Abstracted BatchHandler function into SalaryPredictor.py, deleted BatchHandler.py
